requirements-simple.txt
# Versão MÍNIMA - apenas pacotes puros Python
Flask==3.0.0
folium==0.15.1
requests==2.31.0
numpy==1.26.3
pandas==2.1.4
python-dotenv==1.0.0
Pillow==10.2.0
